requirements.txt
faiss_cpu==1.9.0.post1
Flask==3.1.0
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==2.2.2
Requests==2.32.3
sentence_transformers==2.6.1
transformers==4.43.2
~ransformers==4.39.1
~ransformers==4.41.1
